feat: hierarchy create with SG support, RBAC propagation, RG-scoped configs
- ResourceGroup flow: Site + Config + ConfigRef (3 resources)
- ServiceGroup flow: recursive SG + Site + Config + ConfigRef (up to 3 levels, 12 resources)
- Configs are always RG-scoped (--resource-group required)
- RBAC propagation wait after each SG creation (polls site list)
- Removed old hierarchy_create.py, using clean v2

E2E tested: 3-level SG hierarchy created successfully.
