+++ src/workload-orchestration/azext_workload_orchestration/_params.py
@@ -90,7 +90,7 @@ def load_arguments(self, _):  # pylint: disable=unused-argument
 
     with self.argument_context('workload-orchestration hierarchy create') as c:
         c.argument('resource_group', options_list=['--resource-group', '-g'],
-                   help='Resource group (required for ResourceGroup type hierarchy).')
+                   help='Resource group for Configuration resources.', required=True)
         c.argument('configuration_location', options_list=['--configuration-location'],
                    help='Azure region for the Configuration resource (e.g., eastus2euap).', required=True)
         c.argument('hierarchy_spec', options_list=['--hierarchy-spec'],
