+++ src/workload-orchestration/azext_workload_orchestration/onboarding/__init__.py
@@ -11,7 +11,7 @@ into single-command operations to reduce onboarding steps.
 
 from azext_workload_orchestration.onboarding.target_prepare import target_prepare
 from azext_workload_orchestration.onboarding.target_deploy import target_deploy as _target_deploy
-from azext_workload_orchestration.onboarding.hierarchy_create_v2 import hierarchy_create as _hierarchy_create
+from azext_workload_orchestration.onboarding.hierarchy_create import hierarchy_create as _hierarchy_create
 
 
 def target_init(
